removing and refactor changes
